requirements.txt
wheel
autopep8
Flask
Flask-RESTful
Flask-SQLAlchemy
SQLAlchemy
Flask-JWT-Extended
mysql-connector-python
flask-mysqldb
pylint

